# apps/api/internal/auditzip/api.gen.go
// Package auditzip provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package auditzip

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for AuditZipJobStatus.
const (
	Canceled  AuditZipJobStatus = "canceled"
	Failed    AuditZipJobStatus = "failed"
	Queued    AuditZipJobStatus = "queued"
	Running   AuditZipJobStatus = "running"
	Succeeded AuditZipJobStatus = "succeeded"
)

// Defines values for AuditZipRequestFormat.
const (
	Zip AuditZipRequestFormat = "zip"
)

// Defines values for ConflictErrorConflictReason.
const (
	DuplicateJob            ConflictErrorConflictReason = "duplicate_job"
	IdempotencyBodyMismatch ConflictErrorConflictReason = "idempotency_body_mismatch"
	IdempotencyReplay       ConflictErrorConflictReason = "idempotency_replay"
	NotCancelable           ConflictErrorConflictReason = "not_cancelable"
)

// AuditZipJob defines model for AuditZipJob.
type AuditZipJob struct {
	// CanCancel true when cancel=true is accepted
	CanCancel *bool `json:"canCancel,omitempty"`

	// CriteriaHash SHA-256 hex hash of the request criteria for audit chain
	CriteriaHash *string            `json:"criteriaHash,omitempty"`
	Error        *InternalError     `json:"error,omitempty"`
	FinishedAt   *time.Time         `json:"finishedAt"`
	JobId        openapi_types.UUID `json:"jobId"`
	Progress     int                `json:"progress"`
	RequestedAt  time.Time          `json:"requestedAt"`
	Result       *AuditZipResult    `json:"result,omitempty"`
	RetryCount   int                `json:"retryCount"`
	StartedAt    *time.Time         `json:"startedAt"`
	Status       AuditZipJobStatus  `json:"status"`
}

// AuditZipJobStatus defines model for AuditZipJob.Status.
type AuditZipJobStatus string

// AuditZipRequest defines model for AuditZipRequest.
type AuditZipRequest struct {
	Format    AuditZipRequestFormat `json:"format"`
	From      openapi_types.Date    `json:"from"`
	MaxAmount *float64              `json:"maxAmount"`
	MinAmount *float64              `json:"minAmount"`
	Partner   *string               `json:"partner"`
	To        openapi_types.Date    `json:"to"`
}

// AuditZipRequestFormat defines model for AuditZipRequest.Format.
type AuditZipRequestFormat string

// AuditZipResult defines model for AuditZipResult.
type AuditZipResult struct {
	// ExpiresAt Expiration timestamp of the signed URL
	ExpiresAt time.Time `json:"expiresAt"`

	// SignedUrl Signed URL valid for configured TTL (default 10 minutes, read-only)
	SignedUrl string `json:"signedUrl"`

	// Size Size in bytes
	Size int `json:"size"`
}

// ConflictError defines model for ConflictError.
type ConflictError struct {
	Code           string                      `json:"code"`
	ConflictReason ConflictErrorConflictReason `json:"conflictReason"`
	CorrId         string                      `json:"corrId"`
	Message        string                      `json:"message"`
	Retryable      bool                        `json:"retryable"`
}

// ConflictErrorConflictReason defines model for ConflictError.ConflictReason.
type ConflictErrorConflictReason string

// ForbiddenError defines model for ForbiddenError.
type ForbiddenError struct {
	Code      string `json:"code"`
	CorrId    string `json:"corrId"`
	Message   string `json:"message"`
	Retryable bool   `json:"retryable"`
}

// InternalError defines model for InternalError.
type InternalError struct {
	Code      string `json:"code"`
	CorrId    string `json:"corrId"`
	Message   string `json:"message"`
	Retryable bool   `json:"retryable"`
}

// NotFoundError defines model for NotFoundError.
type NotFoundError struct {
	Code      string `json:"code"`
	CorrId    string `json:"corrId"`
	Message   string `json:"message"`
	Retryable bool   `json:"retryable"`
}

// RateLimitError defines model for RateLimitError.
type RateLimitError struct {
	Code    string `json:"code"`
	CorrId  string `json:"corrId"`
	Message string `json:"message"`

	// RetryAfterSeconds Minimum seconds to wait before retry
	RetryAfterSeconds int  `json:"retryAfterSeconds"`
	Retryable         bool `json:"retryable"`
}

// RequestTooLargeError defines model for RequestTooLargeError.
type RequestTooLargeError struct {
	Code      string    `json:"code"`
	CorrId    string    `json:"corrId"`
	Message   string    `json:"message"`
	Retryable bool      `json:"retryable"`
	SplitHint SplitHint `json:"splitHint"`
}

// SplitHint defines model for SplitHint.
type SplitHint struct {
	// ApproxSizeMB Approximate size per chunk in MB
	ApproxSizeMB float64 `json:"approxSizeMB"`

	// Chunks Suggested number of chunks
	Chunks int `json:"chunks"`
}

// ValidationError defines model for ValidationError.
type ValidationError struct {
	Code      string                `json:"code"`
	CorrId    string                `json:"corrId"`
	Errors    []ValidationErrorItem `json:"errors"`
	Message   string                `json:"message"`
	Retryable bool                  `json:"retryable"`
}

// ValidationErrorItem defines model for ValidationErrorItem.
type ValidationErrorItem struct {
	Code    string `json:"code"`
	Message string `json:"message"`
	Path    string `json:"path"`
}

// CorrelationId defines model for CorrelationId.
type CorrelationId = openapi_types.UUID

// IdempotencyKey defines model for IdempotencyKey.
type IdempotencyKey = openapi_types.UUID

// TenantId defines model for TenantId.
type TenantId = string

// AuditJobAccepted defines model for AuditJobAccepted.
type AuditJobAccepted = AuditZipJob

// AuditJobStatus defines model for AuditJobStatus.
type AuditJobStatus = AuditZipJob

// Conflict defines model for Conflict.
type Conflict = ConflictError

// Forbidden defines model for Forbidden.
type Forbidden = ForbiddenError

// NotFound defines model for NotFound.
type NotFound = NotFoundError

// RateLimit defines model for RateLimit.
type RateLimit = RateLimitError

// RequestTooLarge defines model for RequestTooLarge.
type RequestTooLarge = RequestTooLargeError

// GetAuditZipJobParams defines parameters for GetAuditZipJob.
type GetAuditZipJobParams struct {
	// Cancel Request cancellation when the job is in running state.
	Cancel *bool `form:"cancel,omitempty" json:"cancel,omitempty"`

	// XCorrelationId Correlation ID for tracing and audit hash chain (echoed back)
	XCorrelationId CorrelationId `json:"X-Correlation-Id"`

	// XTenantId Tenant identifier for RBAC and storage segregation
	XTenantId TenantId `json:"X-Tenant-Id"`
}

// EnqueueAuditZipParams defines parameters for EnqueueAuditZip.
type EnqueueAuditZipParams struct {
	// XCorrelationId Correlation ID for tracing and audit hash chain (echoed back)
	XCorrelationId CorrelationId `json:"X-Correlation-Id"`

	// XTenantId Tenant identifier for RBAC and storage segregation
	XTenantId TenantId `json:"X-Tenant-Id"`

	// IdempotencyKey Required idempotency key (UUID). Same key + same body returns the same job. Same key + different body returns 409 conflictReason=idempotency_body_mismatch.
	IdempotencyKey IdempotencyKey `json:"Idempotency-Key"`
}

// EnqueueAuditZipJSONRequestBody defines body for EnqueueAuditZip for application/json ContentType.
type EnqueueAuditZipJSONRequestBody = AuditZipRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get audit ZIP job status
	// (GET /audit/jobs/{jobId})
	GetAuditZipJob(w http.ResponseWriter, r *http.Request, jobId openapi_types.UUID, params GetAuditZipJobParams)
	// Enqueue audit ZIP export job
	// (POST /audit/zip)
	EnqueueAuditZip(w http.ResponseWriter, r *http.Request, params EnqueueAuditZipParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get audit ZIP job status
// (GET /audit/jobs/{jobId})
func (_ Unimplemented) GetAuditZipJob(w http.ResponseWriter, r *http.Request, jobId openapi_types.UUID, params GetAuditZipJobParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Enqueue audit ZIP export job
// (POST /audit/zip)
func (_ Unimplemented) EnqueueAuditZip(w http.ResponseWriter, r *http.Request, params EnqueueAuditZipParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAuditZipJob operation middleware
func (siw *ServerInterfaceWrapper) GetAuditZipJob(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "jobId" -------------
	var jobId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "jobId", chi.URLParam(r, "jobId"), &jobId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "jobId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAuditZipJobParams

	// ------------- Optional query parameter "cancel" -------------

	err = runtime.BindQueryParameter("form", true, false, "cancel", r.URL.Query(), &params.Cancel)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cancel", Err: err})
		return
	}

	headers := r.Header

	// ------------- Required header parameter "X-Correlation-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Correlation-Id")]; found {
		var XCorrelationId CorrelationId
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Correlation-Id", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Correlation-Id", valueList[0], &XCorrelationId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Correlation-Id", Err: err})
			return
		}

		params.XCorrelationId = XCorrelationId

	} else {
		err := fmt.Errorf("Header parameter X-Correlation-Id is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "X-Correlation-Id", Err: err})
		return
	}

	// ------------- Required header parameter "X-Tenant-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Tenant-Id")]; found {
		var XTenantId TenantId
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Tenant-Id", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Tenant-Id", valueList[0], &XTenantId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Tenant-Id", Err: err})
			return
		}

		params.XTenantId = XTenantId

	} else {
		err := fmt.Errorf("Header parameter X-Tenant-Id is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "X-Tenant-Id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAuditZipJob(w, r, jobId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// EnqueueAuditZip operation middleware
func (siw *ServerInterfaceWrapper) EnqueueAuditZip(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params EnqueueAuditZipParams

	headers := r.Header

	// ------------- Required header parameter "X-Correlation-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Correlation-Id")]; found {
		var XCorrelationId CorrelationId
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Correlation-Id", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Correlation-Id", valueList[0], &XCorrelationId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Correlation-Id", Err: err})
			return
		}

		params.XCorrelationId = XCorrelationId

	} else {
		err := fmt.Errorf("Header parameter X-Correlation-Id is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "X-Correlation-Id", Err: err})
		return
	}

	// ------------- Required header parameter "X-Tenant-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Tenant-Id")]; found {
		var XTenantId TenantId
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Tenant-Id", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Tenant-Id", valueList[0], &XTenantId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Tenant-Id", Err: err})
			return
		}

		params.XTenantId = XTenantId

	} else {
		err := fmt.Errorf("Header parameter X-Tenant-Id is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "X-Tenant-Id", Err: err})
		return
	}

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey IdempotencyKey
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err := fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.EnqueueAuditZip(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/audit/jobs/{jobId}", wrapper.GetAuditZipJob)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/audit/zip", wrapper.EnqueueAuditZip)
	})

	return r
}